django/requirements.txt
django<4
# Auth Related
djangorestframework<4
djangorestframework-simplejwt<5
django-rest-passwordreset<2
# Admin
django-import-export<3
# DB / Storage / Server
psycopg2<3
gunicorn<21
django-cors-headers<4
whitenoise<6
requests<3
uuid<2